backend/requirements.txt
# FastAPI
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# NumPy 2.x only where LangChain allows it (3.13+); 3.11 needs numpy<2 per langchain 0.3.x
numpy>=2.0.0,<3; python_version >= "3.13"

# LangChain + LangGraph
langchain==0.3.13
langchain-core==0.3.63
langchain-text-splitters==0.3.8
langchain-openai==0.2.14
langchain-community==0.3.13
langgraph==0.2.60

# Vector store
chromadb==1.5.7

# Hybrid RAG - BM25
rank-bm25==0.2.2

# Document loaders
pypdf>=6.9.2,<7
unstructured[xlsx]==0.18.32
openpyxl==3.1.5
lxml>=6.0.0

# Web search
tavily-python==0.5.0

# Export
python-docx==1.1.2
markdown==3.7
weasyprint==62.3

# Config
python-dotenv==1.0.1

# HTTP client
httpx==0.28.1
